+++ src/main/java/twilightforest/item/ItemBlockTFPlant.java
@@ -118,10 +118,7 @@ public class ItemBlockTFPlant extends ItemBlock {
                         plantBlock.onBlockPlacedBy(world, x, y, z, player, itemStack);
                     }
 
-                    world.playSoundEffect(
-                            (double) (x + 0.5F),
-                            (double) (y + 0.5F),
-                            (double) (z + 0.5F),
+                    world.playSoundEffect(x + 0.5F, y + 0.5F, z + 0.5F,
                             plantBlock.stepSound.getBreakSound(),
                             (plantBlock.stepSound.getVolume() + 1.0F) / 2.0F,
                             plantBlock.stepSound.getPitch() * 0.8F);
