+++ src/main/java/twilightforest/item/ItemTFMoonwormQueen.java
@@ -98,10 +98,7 @@ public class ItemTFMoonwormQueen extends ItemTF {
                     TFBlocks.moonworm.onBlockPlacedBy(world, x, y, z, player, par1ItemStack);
                 }
 
-                world.playSoundEffect(
-                        (double) (x + 0.5F),
-                        (double) (y + 0.5F),
-                        (double) (z + 0.5F),
+                world.playSoundEffect(x + 0.5F, y + 0.5F, z + 0.5F,
                         this.getSound(),
                         TFBlocks.moonworm.stepSound.getVolume() / 2.0F,
                         TFBlocks.moonworm.stepSound.getPitch() * 0.8F);
